requirements.txt
torch
torchvision
transformers
datasets
evaluate
jiwer
pillow
tqdm
accelerate
sentencepiece
pandas
numpy
tensorboard
scikit-learn
